pc - update stories for storybook 8
