+++ frontend/src/stories/pages/CourseOverTime/CourseOverTimeBuildingsIndexPage.stories.js
